+++ src/Overwurd.Model/Services/JwtRefreshTokenProvider.cs
@@ -23,7 +23,7 @@ namespace Overwurd.Model.Services
             await dbContext.SaveChangesAsync(cancellationToken);
         }
 
-        public async Task<JwtRefreshToken> GetUserTokenAsync(long userId, CancellationToken cancellationToken)
+        public async Task<JwtRefreshToken> GetUserTokenAsync(int userId, CancellationToken cancellationToken)
         {
             return await dbContext.JwtRefreshTokens
                                   .Where(x => x.UserId == userId)
@@ -31,7 +31,7 @@ namespace Overwurd.Model.Services
                                   .SingleOrDefaultAsync(cancellationToken);
         }
 
-        public async Task RemoveUserTokenAsync(long userId, CancellationToken cancellationToken)
+        public async Task RemoveUserTokenAsync(int userId, CancellationToken cancellationToken)
         {
             var userToken = await dbContext.JwtRefreshTokens
                                            .FindAsync(new object[] { userId }, cancellationToken: cancellationToken);
